Move state management to browserstate.ts
